tested the BigDecimal class test

+++ src/pt/isel/ccd/arithmetic/CodedData.java
@@ -1,18 +1,20 @@
 package pt.isel.ccd.arithmetic;
 
+import java.math.BigDecimal;
+
 /**
  * Created by trinkes on 03/12/2016.
  */
 public class CodedData {
     private int length;
-    private Double tag;
+    private BigDecimal tag;
 
-    public CodedData(double tag, int length) {
+    public CodedData(BigDecimal tag, int length) {
         this.tag = tag;
         this.length = length;
     }
 
-    public double getTag() {
+    public BigDecimal getTag() {
         return tag;
     }
